Remove multiple_priorities_long_idle_name and multiple_priorities_preemption_disable
