+++ FreeRTOS/Test/CMock/smp/Makefile
@@ -8,11 +8,14 @@ include ../makefile.in
 SUITES	+=	single_priority_no_timeslice
 SUITES	+=	single_priority_timeslice
 SUITES	+=	multiple_priorities_no_timeslice
-SUITES	+=	multiple_priorities_no_timeslice_mock
 SUITES	+=	multiple_priorities_timeslice
+
+# mock SUITES for coverage test
+SUITES	+=	multiple_priorities_no_timeslice_mock
+SUITES	+=	multiple_priorities_long_idle_name_mock
+
+# config assert SUITES
 SUITES	+=	config_assert
-SUITES	+=	multiple_priorities_preemption_disable
-SUITES	+=	multiple_priorities_long_idle_name
 
 # PROJECT and SUITE variables are determined based on path like so:
 #   $(UT_ROOT_DIR)/$(PROJECT)/$(SUITE)
