+++ FreeRTOS/Test/CMock/smp/multiple_priorities_long_idle_name_mock/Makefile
@@ -17,7 +17,7 @@ PROJECT_SRC         :=  tasks.c
 PROJECT_HEADER_DEPS :=  FreeRTOS.h
 
 # SUITE_UT_SRC: .c files that contain test cases (must end in _utest.c)
-SUITE_UT_SRC        :=  multiple_priorities_preemption_disable_utest.c
+SUITE_UT_SRC        :=  covg_multiple_priorities_long_idle_name_utest.c
 
 # SUITE_SUPPORT_SRC: .c files used for testing that do not contain test cases.
 # Paths are relative to PROJECT_DIR
