+++ FreeRTOS/Test/CMock/smp/multiple_priorities_no_timeslice_mock/covg_multiple_priorities_no_timeslice_mock_utest.c
@@ -23,54 +23,42 @@
  * https://github.com/FreeRTOS
  *
  */
-/*! @file config_assert_utest.c */
+/*! @file multiple_priorities_no_timeslice_utest.c */
 
 /* C runtime includes. */
 #include <stdlib.h>
 #include <stdbool.h>
 #include <stdint.h>
 #include <string.h>
-#include <stdio.h>
 
-/* Tasks includes */
+/* Tasl includes */
 #include "FreeRTOS.h"
 #include "FreeRTOSConfig.h"
-#include "../global_vars.h"
-#include "task.h"
-
-/* #include "fake_port.h" */
-#include "portmacro.h"
+#include "event_groups.h"
+#include "queue.h"
 
 /* Test includes. */
 #include "unity.h"
 #include "unity_memory.h"
-#include "CException.h"
-
-
-/* Local includes. */
-/*#include "../smp_utest_common.h" */
+#include "../global_vars.h"
+#include "../smp_utest_common.h"
+#include <assert.h>
 
 /* Mock includes. */
 #include "mock_timers.h"
-#include "mock_list.h"
-#include "mock_list_macros.h"
 #include "mock_fake_assert.h"
 #include "mock_fake_port.h"
+#include "mock_list.h"
+#include "mock_list_macros.h"
 #include "mock_local_portable.h"
 
 
 /* =================================  MACROS  =============================== */
-
 /**
  * @brief CException code for when a configASSERT should be intercepted.
  */
 #define configASSERT_E    0xAA101
 #define EXIT_LOOP         0xAA102
-
-/**
- * @brief simulate up to 10 tasks: add more if needed
- * */
-#define TCB_ARRAY         10
 
 /**
  * @brief Expect a configASSERT from the function called.
@@ -94,7 +82,6 @@
         }                                           \
     } while( 0 )
 
-
 /* ============================  GLOBAL VARIABLES =========================== */
 
 /**
@@ -108,23 +95,28 @@ static int assertionFailed = 1;
 static BaseType_t shouldAbortOnAssertion;
 
 /* ===========================  EXTERN VARIABLES  =========================== */
-extern void vTaskEnterCritical( void );
-extern void vTaskExitCritical( void );
-extern void vTaskExitCriticalFromISR( UBaseType_t uxSavedInterruptStatus );
 
 extern volatile UBaseType_t uxDeletedTasksWaitingCleanUp;
-extern volatile BaseType_t xSchedulerRunning;
-extern volatile UBaseType_t uxSchedulerSuspended;
-extern TCB_t * volatile pxCurrentTCBs[ configNUMBER_OF_CORES ];
+extern volatile UBaseType_t xSchedulerRunning;
 extern volatile BaseType_t xYieldPendings[ configNUMBER_OF_CORES ];
+extern volatile TCB_t * pxCurrentTCBs[ configNUMBER_OF_CORES ];
+extern TaskHandle_t xIdleTaskHandles[configNUMBER_OF_CORES];
+extern volatile BaseType_t xYieldForTask;
+extern volatile BaseType_t xYieldRequired;
+extern volatile UBaseType_t uxCurrentNumberOfTasks;
+extern volatile UBaseType_t uxSchedulerSuspended;
 extern volatile UBaseType_t uxTopReadyPriority;
 extern List_t pxReadyTasksLists[ configMAX_PRIORITIES ];
 extern UBaseType_t uxTaskNumber;
 extern volatile TickType_t xTickCount;
 extern volatile TickType_t xNextTaskUnblockTime;
-extern TaskHandle_t xIdleTaskHandles[ configNUMBER_OF_CORES ];
 
-/* ===========================  STATIC FUNCTIONS  =========================== */
+/* ===========================  EXTERN FUNCTIONS  =========================== */
+
+extern BaseType_t prvCreateIdleTasks( void );
+
+/* ==========================  STATIC FUNCTIONS  ========================== */
+
 void vApplicationMinimalIdleHook( void )
 {
 }
@@ -145,26 +137,18 @@ static void vFakeAssertStub( bool x,
     }
 }
 
-void * pvPortMalloc( size_t xSize )
-{
-    return unity_malloc( xSize );
-}
-
-void vPortFree( void * pv )
-{
-    return unity_free( pv );
-}
 /* ============================  Unity Fixtures  ============================ */
 /*! called before each testcase */
 void setUp( void )
 {
-    uxDeletedTasksWaitingCleanUp;
     vFakeAssert_StubWithCallback( vFakeAssertStub );
+    UnityMalloc_StartTest();
 }
 
 /*! called after each testcase */
 void tearDown( void )
 {
+    UnityMalloc_EndTest();
 }
 
 /*! called at the beginning of the whole suite */
@@ -178,7 +162,225 @@ int suiteTearDown( int numFailures )
     return numFailures;
 }
 
+
+/* ==========================  HELPER FUNCTIONS  ========================== */
+
+void * pvPortMalloc( size_t xSize )
+{
+    return unity_malloc( xSize );
+}
+
+void vPortFree( void * pv )
+{
+    return unity_free( pv );
+}
+
 /* ==============================  Test Cases  ============================== */
+
+/**
+ * @brief vTaskSuspend - suspends the running task
+ *
+ * This test is used to suspend a running task when task is actively
+ * running and not scheduled to yield.
+ *
+ * <b>Coverage</b>
+ * @code{c}
+ * if( xSchedulerRunning != pdFALSE )
+ * {
+ *      if( xTaskRunningOnCore == portGET_CORE_ID() )
+ *      {
+ *          configASSERT( uxSchedulerSuspended == 0 );
+ *          vTaskYieldWithinAPI();
+ *      }
+ *      else
+ *      {
+ *          prvYieldCore( xTaskRunningOnCore );
+ *      }
+ * }
+ * else
+ * {
+ *      mtCOVERAGE_TEST_MARKER();
+ * }
+ * @endcode
+ * ( xSchedulerRunning != pdFALSE ) is false.
+ */
+void test_vTaskSuspend_scheduler_running_false( void )
+{
+    TCB_t xTaskTCBs[ 1 ] = { NULL };
+
+    /* Setup the variables and structure. */
+    xTaskTCBs[ 0 ].uxPriority = 1;
+    xYieldPendings[ 0 ] = pdFALSE;
+    pxCurrentTCBs[ 0 ] = &xTaskTCBs[ 0 ];
+    pxCurrentTCBs[ 0 ]->xTaskRunState = 1;
+    xSchedulerRunning = pdFALSE;
+
+    vFakePortEnterCriticalSection_Expect();
+    uxListRemove_ExpectAnyArgsAndReturn( pdTRUE );
+    listLIST_ITEM_CONTAINER_ExpectAnyArgsAndReturn( NULL );
+    vListInsertEnd_ExpectAnyArgs();
+    vFakePortExitCriticalSection_Expect();
+
+    /* API call. */
+    vTaskSuspend( &xTaskTCBs[ 0 ] );
+
+    /* Validation. */
+    TEST_ASSERT_EQUAL( pdFALSE, xYieldPendings[ 0 ] );
+    TEST_ASSERT_EQUAL( 1, pxCurrentTCBs[ 0 ]->xTaskRunState );
+}
+
+/**
+ * @brief vTaskSuspend - suspends the task
+ *
+ * This test is used to ensure that one of the macro's conditions
+ * where TaskRunState is greater than zero is set to false.
+ *
+ * <b>Coverage</b>
+ * @code{c}
+ * if( taskTASK_IS_RUNNING( pxTCB ) == pdTRUE )
+ * @endcode
+ * ( taskTASK_IS_RUNNING( pxTCB ) ) is false.
+ */
+
+void test_vTaskSuspend_running_state_below_range( void )
+{
+    TCB_t xTaskTCBs[ 1 ] = { NULL };
+
+    /* Setup the variables and structure. */
+    xTaskTCBs[ 0 ].uxPriority = 1;
+    xYieldPendings[ 0 ] = pdFALSE;
+    pxCurrentTCBs[ 0 ] = &xTaskTCBs[ 0 ];
+    pxCurrentTCBs[ 0 ]->xTaskRunState = -1;
+    xSchedulerRunning = pdFALSE;
+
+    vFakePortEnterCriticalSection_Expect();
+    uxListRemove_ExpectAnyArgsAndReturn( pdTRUE );
+    listLIST_ITEM_CONTAINER_ExpectAnyArgsAndReturn( NULL );
+    vListInsertEnd_ExpectAnyArgs();
+    vFakePortExitCriticalSection_Expect();
+
+    /* API call. */
+    vTaskSuspend( &xTaskTCBs[ 0 ] );
+
+    /* Validation. */
+    TEST_ASSERT_EQUAL( pdFALSE, xYieldPendings[ 0 ] );
+    TEST_ASSERT_EQUAL( -1, pxCurrentTCBs[ 0 ]->xTaskRunState );
+}
+
+/**
+ * @brief vTaskSuspend - suspends the task
+ *
+ * This test is used to cover the case where the other macro condition where
+ * TaskRunState is less than configNUMBER_OF_CORES is set to false.
+ *
+ * <b>Coverage</b>
+ * @code{c}
+ * if( taskTASK_IS_RUNNING( pxTCB ) == pdTRUE )
+ * @endcode
+ * ( taskTASK_IS_RUNNING( pxTCB ) ) is false.
+ */
+
+void test_vTaskSuspend_running_state_above_range( void )
+{
+    TCB_t xTaskTCBs[ 1 ] = { NULL };
+
+    /* Setup the variables and structure. */
+    xTaskTCBs[ 0 ].uxPriority = 1;
+    xYieldPendings[ 0 ] = pdFALSE;
+    pxCurrentTCBs[ 0 ] = &xTaskTCBs[ 0 ];
+    pxCurrentTCBs[ 0 ]->xTaskRunState = configNUMBER_OF_CORES + 1;
+
+    vFakePortEnterCriticalSection_Expect();
+    uxListRemove_ExpectAnyArgsAndReturn( pdTRUE );
+    listLIST_ITEM_CONTAINER_ExpectAnyArgsAndReturn( NULL );
+    vListInsertEnd_ExpectAnyArgs();
+    vFakePortExitCriticalSection_Expect();
+
+    /* API call. */
+    vTaskSuspend( &xTaskTCBs[ 0 ] );
+
+    /* Validation. */
+    TEST_ASSERT_EQUAL( pdFALSE, xYieldPendings[ 0 ] );
+    TEST_ASSERT_EQUAL( 17, pxCurrentTCBs[ 0 ]->xTaskRunState );
+}
+
+/**
+ * @brief vTaskPrioritySet - sets the priority
+ *
+ * This test is to change the priorities of non
+ * running tasks
+ *
+ * <b>Coverage</b>
+ * @code{c}
+ * else if( taskTASK_IS_RUNNING( pxTCB ) == pdTRUE )
+ * @endcode
+ * ( taskTASK_IS_RUNNING( pxTCB ) ) is false.
+ */
+void test_vTaskPrioritySet_non_running_state( void )
+{
+    TCB_t xTaskTCBs[ 1 ] = { NULL };
+
+    /* Setup the variables and structure. */
+    xTaskTCBs[ 0 ].uxPriority = 4;
+    xTaskTCBs[ 0 ].uxBasePriority = 4;
+    xTaskTCBs[ 0 ].xTaskRunState = configNUMBER_OF_CORES + 1;
+
+    vFakeAssert_Ignore();
+    vFakePortCheckIfInISR_StopIgnore();
+    vFakePortEnterCriticalSection_Ignore();
+    vFakePortYieldCore_CMockIgnore();
+    listGET_LIST_ITEM_VALUE_ExpectAnyArgsAndReturn( ( TickType_t ) 0x80000000UL );
+    listIS_CONTAINED_WITHIN_ExpectAnyArgsAndReturn( pdFALSE );
+    vFakePortExitCriticalSection_Ignore();
+
+    /* API call. */
+    vTaskPrioritySet( &xTaskTCBs[ 0 ], 2 );
+
+    /* Validation. */
+    TEST_ASSERT_EQUAL( 17, xTaskTCBs[ 0 ].xTaskRunState );
+    TEST_ASSERT_EQUAL( 2, xTaskTCBs[ 0 ].uxPriority );
+}
+
+/**
+ * @brief vTaskPrioritySet - sets the priority
+ *
+ * This test verifies that the current task is not
+ * preempted by any other task of higher priority.
+ *
+ * <b>Coverage</b>
+ * @code{c}
+ * if( pxTCB->xPreemptionDisable == pdFALSE )
+ * @endcode
+ * ( pxTCB->xPreemptionDisable == pdFALSE ) is false.
+ */
+void test_vTaskPrioritySet_running_state( void )
+{
+    TCB_t xTaskTCBs[ 1 ] = { NULL };
+
+    /* Setup the variables and structure. */
+    xTaskTCBs[ 0 ].uxPriority = 4;
+    xTaskTCBs[ 0 ].uxBasePriority = 4;
+    xTaskTCBs[ 0 ].xPreemptionDisable = pdTRUE;
+    xTaskTCBs[ 0 ].xTaskRunState = 1;
+
+    BaseType_t xYieldRequired = pdFALSE;
+    BaseType_t xYieldForTask = pdFALSE;
+
+    vFakeAssert_Ignore();
+    vFakePortCheckIfInISR_StopIgnore();
+    vFakePortEnterCriticalSection_Ignore();
+    vFakePortYieldCore_CMockIgnore();
+    listGET_LIST_ITEM_VALUE_ExpectAnyArgsAndReturn( ( TickType_t ) 0x80000000UL );
+    listIS_CONTAINED_WITHIN_ExpectAnyArgsAndReturn( pdFALSE );
+    vFakePortExitCriticalSection_Ignore();
+
+    /* API call. */
+    vTaskPrioritySet( &xTaskTCBs[ 0 ], 2 );
+
+    /* Validation. */
+    TEST_ASSERT_EQUAL( pdFALSE, xYieldRequired );
+    TEST_ASSERT_EQUAL( pdFALSE, xYieldForTask );
+}
 
 /**
  * @brief prvYieldCore - xCoreID is not equal to current core id.
@@ -451,7 +653,6 @@ void test_vTaskPreemptionDisable_null_handle( void )
     /* Test Verifications */
     TEST_ASSERT_EQUAL( pdTRUE, pxCurrentTCBs[ 0 ]->xPreemptionDisable );
 }
-
 
 /**
  * @brief This test ensures that when we call vTaskSuspendAll and we task of the
@@ -939,6 +1140,14 @@ void test_prvCreateIdleTasks_name_too_long( void )
     /* Test Verifications */
     xIdleTask = ( TCB_t * ) xIdleTaskHandles[ 0 ];
     TEST_ASSERT_EQUAL_STRING( configIDLE_TASK_NAME, xIdleTask->pcTaskName );
+
+    /* Clean up idle task. */
+    for( i = 0; i < configNUMBER_OF_CORES; i++ )
+    {
+        vPortFreeStack( xIdleTaskHandles[ i ]->pxStack );
+        vPortFree( xIdleTaskHandles[ i ] );
+        xIdleTaskHandles[ i ] = NULL;
+    }
 }
 
 /**
