+++ FreeRTOS/Test/CMock/smp/multiple_priorities_no_timeslice_mock/local_portable.h
@@ -1,10 +1,12 @@
-#ifndef PORTABLE_H
-#define PORTABLE_H
-void * pvPortMalloc( size_t xSize );
-void vPortFree( void * pv ) ;
+#ifndef LOCAL_PORTABLE_H
+#define LOCAL_PORTABLE_H
+
 StackType_t * pxPortInitialiseStack( StackType_t * pxTopOfStack,
-                                             TaskFunction_t pxCode,
-                                             void * pvParameters ) ;
-BaseType_t xPortStartScheduler( void ) ;
-void vPortEndScheduler( void ) ;
-#endif 
+                                     TaskFunction_t pxCode,
+                                     void * pvParameters );
+
+BaseType_t xPortStartScheduler( void );
+
+void vPortEndScheduler( void );
+
+#endif
